+++ common/src/comp/agent.rs
@@ -367,6 +367,7 @@ impl<'a> From<&'a Body> for Psyche {
                 },
                 Body::BipedSmall(biped_small) => match biped_small.species {
                     biped_small::Species::Gnarling => 0.2,
+                    biped_small::Species::Mandragora => 0.1,
                     biped_small::Species::Adlet => 0.2,
                     biped_small::Species::Haniwa => 0.1,
                     biped_small::Species::Sahagin => 0.1,
@@ -379,7 +380,8 @@ impl<'a> From<&'a Body> for Psyche {
                     | biped_small::Species::Irrwurz
                     | biped_small::Species::ShamanicSpirit
                     | biped_small::Species::Jiangshi
-                    | biped_small::Species::Bushly => 0.0,
+                    | biped_small::Species::Bushly
+                    | biped_small::Species::GnarlingChieftain => 0.0,
 
                     _ => 0.5,
                 },