+++ server/agent/src/action_nodes.rs
@@ -1204,9 +1204,15 @@ impl<'a> AgentData<'a> {
         // Wield the weapon as running towards the target
         controller.push_action(ControlAction::Wield);
 
-        let min_attack_dist = (self.body.map_or(0.5, |b| b.max_radius()) + DEFAULT_ATTACK_RANGE)
-            * self.scale
-            + tgt_data.body.map_or(0.5, |b| b.max_radius()) * tgt_data.scale.map_or(1.0, |s| s.0);
+        // Information for attack checks
+        // 'min_attack_dist' uses DEFAULT_ATTACK_RANGE, while 'body_dist' does not
+        let self_radius = self.body.map_or(0.5, |b| b.max_radius()) * self.scale;
+        let self_attack_range =
+            (self.body.map_or(0.5, |b| b.max_radius()) + DEFAULT_ATTACK_RANGE) * self.scale;
+        let tgt_radius =
+            tgt_data.body.map_or(0.5, |b| b.max_radius()) * tgt_data.scale.map_or(1.0, |s| s.0);
+        let min_attack_dist = self_attack_range + tgt_radius;
+        let body_dist = self_radius + tgt_radius;
         let dist_sqrd = self.pos.0.distance_squared(tgt_data.pos.0);
         let angle = self
             .ori
@@ -1358,6 +1364,7 @@ impl<'a> AgentData<'a> {
         }
 
         let attack_data = AttackData {
+            body_dist,
             min_attack_dist,
             dist_sqrd,
             angle,
@@ -1617,9 +1624,14 @@ impl<'a> AgentData<'a> {
             Tactic::Yeti => {
                 self.handle_yeti_attack(agent, controller, &attack_data, tgt_data, read_data)
             },
-            Tactic::Harvester => {
-                self.handle_harvester_attack(agent, controller, &attack_data, tgt_data, read_data)
-            },
+            Tactic::Harvester => self.handle_harvester_attack(
+                agent,
+                controller,
+                &attack_data,
+                tgt_data,
+                read_data,
+                rng,
+            ),
             Tactic::Cardinal => self.handle_cardinal_attack(
                 agent,
                 controller,
@@ -1673,7 +1685,7 @@ impl<'a> AgentData<'a> {
                 self.handle_mandragora(agent, controller, &attack_data, tgt_data, read_data)
             },
             Tactic::WoodGolem => {
-                self.handle_wood_golem(agent, controller, &attack_data, tgt_data, read_data)
+                self.handle_wood_golem(agent, controller, &attack_data, tgt_data, read_data, rng)
             },
             Tactic::GnarlingChieftain => self.handle_gnarling_chieftain(
                 agent,